% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_dimred.R
\name{theme_dimred}
\alias{theme_dimred}
\title{Brandon's Theme for Dimensionality Reduction Plots}
\usage{
theme_dimred(label = TRUE, base_size = 8, text_size = 7, square = TRUE)
}
\arguments{
\item{label}{Logical. If TRUE (default), show axis labels and lines; otherwise, hide them.}

\item{square}{Logical. If TRUE (default), sets aspect ratio to 1 (square plot); if FALSE, does not set aspect ratio.}
}
\value{
A ggplot2 theme or list of theme + guides.
}
\description{
A minimalist theme optimized for UMAP, PCA, t-SNE, and similar plots,
with optional axis labeling.
}
\examples{
library(ggplot2)
df <- data.frame(UMAP1 = rnorm(100), UMAP2 = rnorm(100))
ggplot(df, aes(UMAP1, UMAP2)) +
  geom_point() +
  theme_dimred()
}
